vp9_mcomp,get_cost_surf_min: quiet conversion warning
visual studio will warn if a 32-bit shift is implicitly converted to 64.
in this case integer storage is enough for the result.
since:
f3a9ae5ba Fix ubsan failure in vp9_mcomp.c.

+++ vp9/encoder/vp9_mcomp.c
@@ -383,7 +383,7 @@ static void get_cost_surf_min(int *cost_list, int *ir, int *ic, int bits) {
   const int64_t y0 = cost_list[1] - 2 * (int64_t)cost_list[0] + cost_list[3];
   const int64_t x1 = (int64_t)cost_list[4] - cost_list[2];
   const int64_t y1 = cost_list[4] - 2 * (int64_t)cost_list[0] + cost_list[2];
-  const int64_t b = 1 << (bits - 1);
+  const int b = 1 << (bits - 1);
   *ic = (int)divide_and_round(x0 * b, y0);
   *ir = (int)divide_and_round(x1 * b, y1);
 }
